tools/requirements.txt
# 网页自动化
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0
selenium>=4.15.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
sqlalchemy>=2.0.0

# AI工具
openai>=1.0.0
ollama>=0.1.0

# 系统工具
python-dotenv>=1.0.0

# 其他
tqdm>=4.65.0
rich>=13.0.0
